requirements.txt
streamlit==1.28.1
pillow==10.0.1
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0